# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reflection Portfolio & Zen Coach API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ZenMessageRole =
  (typeof ZenMessageRole)[keyof typeof ZenMessageRole];

export const ZenMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ZenMessage {
  role: ZenMessageRole;
  content: string;
}

export interface ZenChatInput {
  messages: ZenMessage[];
}

export interface ZenReply {
  reply: string;
}

export interface ZenScheduleInput {
  /** The user's mental health concern or current emotional state */
  concern: string;
}

export type ZenScheduleItemCategory =
  (typeof ZenScheduleItemCategory)[keyof typeof ZenScheduleItemCategory];

export const ZenScheduleItemCategory = {
  meditation: "meditation",
  journaling: "journaling",
  affirmation: "affirmation",
  movement: "movement",
  rest: "rest",
  social: "social",
  creative: "creative",
} as const;

export interface ZenScheduleItem {
  time: string;
  activity: string;
  duration: string;
  category: ZenScheduleItemCategory;
  description?: string;
}

export interface ZenSchedule {
  items: ZenScheduleItem[];
  message: string;
}