# lib/api-zod/src/generated/types/zenSchedule.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reflection Portfolio & Zen Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { ZenScheduleItem } from "./zenScheduleItem";

export interface ZenSchedule {
  items: ZenScheduleItem[];
  message: string;
}