# lib/api-zod/src/generated/types/zenScheduleInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reflection Portfolio & Zen Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface ZenScheduleInput {
  /** The user's mental health concern or current emotional state */
  concern: string;
}