# lib/api-zod/src/generated/types/zenScheduleItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reflection Portfolio & Zen Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { ZenScheduleItemCategory } from "./zenScheduleItemCategory";

export interface ZenScheduleItem {
  time: string;
  activity: string;
  duration: string;
  category: ZenScheduleItemCategory;
  description?: string;
}